Extracts Profile object to controller and model
