+++ src/controllers/ProfileController.js
@@ -1,0 +1,28 @@
+const Profile = require("../models/Profile");
+
+module.exports = {
+    index(request, response) {
+        return response.render("profile", { profile: Profile.get() });
+    },
+    update(request, response) {
+        const data = request.body;
+
+        const weeksPerYear = 52;
+
+        const weeksPerMonth = (weeksPerYear - data["vacation-per-year"]) / 12;
+
+        const weekTotalHours = data["hours-per-day"] * data["days-per-week"];
+
+        const monthlyTotalHours = weeksPerMonth  * weekTotalHours;
+
+        const valueHour = data["monthly-budget"] / monthlyTotalHours;
+
+        Profile.update({
+            ...Profile.get(),
+            ...request.body,
+            "value-hour": valueHour
+        });
+        
+        return response.redirect("/profile");
+    }
+};
